requirements.txt
brewer2mpl==1.4.1
cycler==0.10.0
Cython==0.29.17
decamelize==0.1.2
ggplot==0.11.5
grpcio==1.33.2
grpcio-tools==1.33.2
joblib==0.14.1
kiwisolver==1.1.0
machine-learning==0.0.1
matplotlib==3.0.3
nose==1.3.7
numpy==1.18.5
pandas==0.25.3
patsy==0.5.1
protobuf==3.13.0
pybind11==2.4.3
pycairo==1.20.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
statsmodels==0.10.2
tornado==6.0.4
